fix(types/primitives): moved chainspec module to types
