+++ crates/irys-reth/src/lib.rs
@@ -62,7 +62,6 @@ use crate::{
     payload_service_builder::IyrsPayloadServiceBuilder,
 };
 
-pub mod chainspec;
 pub mod evm;
 pub mod payload;
 pub mod payload_builder_builder;
@@ -70,7 +69,7 @@ pub mod payload_service_builder;
 pub mod pd_tx;
 pub mod precompile;
 pub mod shadow_tx;
-pub use chainspec::{IrysChainHardforks, IrysHardfork};
+pub use irys_types::chainspec::{IrysChainHardforks, IrysHardfork};
 pub use shadow_tx::{IRYS_SHADOW_EXEC, SHADOW_TX_DESTINATION_ADDR};
 
 #[must_use]
