+++ crates/types/Cargo.toml
@@ -44,6 +44,7 @@ toml.workspace = true
 alloy-primitives.workspace = true
 reth-codecs.workspace = true
 reth-primitives.workspace = true
+reth-primitives-traits.workspace = true
 reth-chainspec.workspace = true
 alloy-eips.workspace = true
 alloy-genesis.workspace = true
